requirements.txt
faiss_cpu==1.7.4
numpy==1.26.3
pandas==2.1.4
Pillow==10.2.0
Requests==2.31.0
torch==2.1.2
torchvision==0.16.2
gradio